swap-api/simple-quote-and-swap-instructions/requirements.txt
requests==2.32.3
base58==2.1.1
solders==0.26.0
python-dotenv==1.0.1
solana==0.36.6